+++ plugin/src/adapter/kotlin/org/jlleitschuh/gradle/ktlint/worker/BaselineReporterAdapter.kt
@@ -1,7 +1,0 @@
-package org.jlleitschuh.gradle.ktlint.worker
-
-interface BaselineReporterAdapter {
-    fun onLintError(file: String, err: SerializableLintError, corrected: Boolean)
-}
-
-interface BaselineReporterAdapterFactory